services/training/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "training"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "xgboost==2.1.0",
    "scikit-learn>=1.5",
    "onnxmltools>=1.12",
    "onnxruntime==1.18.1",
    "mlflow>=2.14,<3.0",
    "pandas>=2.2",
    "numpy>=1.26,<2.0",
    "pyiceberg[s3]==0.7.0",
    "pyarrow>=16.0,<18.0",
    "feast[redis]==0.40.0",
    "scipy>=1.13",
    "boto3>=1.34",
    "python-dotenv>=1.0",
    "structlog>=24.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "jupyter>=1.0", "matplotlib>=3.9", "seaborn>=0.13"]

[project.scripts]
train = "training.pipeline:main"

[tool.hatch.build.targets.wheel]
packages = ["src/training"]
